fix missing tilt dependency
was accidentally removed in this commit:

https://github.com/moove-it/sidekiq-scheduler/commit/e173071d8fb6ef7e767
fee0d6953c69449c257dd#diff-6706bb7d3bf023a73d5dd7443cb89459L25

but there are still dependencies on tilt in the code base.  If the
scheduler is a part of a rails project, tilt is already included; but
if the scheduler is not part of a rails project, then an `LoadError`
can occur.